+++ execdriver/native/default_template.go
@@ -1,4 +1,4 @@
-package docker
+package native
 
 import (
 	"github.com/dotcloud/docker/pkg/cgroups"
